fix(score): 手键结算按小数毫秒解析原始时长
为什么：客户端逐符时长是 performance.now() 差值（带小数），上传侧 measurePage 本来就按
Double 校验，结算侧却按 Integer 解，Gson 抛 NumberFormatException（Expected an int but
was 73.20000000111759），整页被判「moresTime/patLogs JSON 损坏，拒绝写入」，finish 返回
202、成绩恒为 0。手键组训与岗位手键共用 MessageComparisonService -> TickerPatUtils 这条
结算路径，两域都受影响。

- TickerPatUtils：moresTime/times 相关 9 处 Integer -> Double；扣分与统计仍按整毫秒，
  取值处就地 Math.round，不改下游语义。
- PostTelegramTrainFinishInfoDto.PatLogs.value：Integer -> Double。

验证：用 keysim 驱动真实页面拍完一页真实报底后，finish 由 202 转为结算成功，学员
isFinish=1、errorNumber=0、服务端重算码率 54 字符/分、成绩页正确率 100%。
全量 mvnw verify 需等 TickerTape 那片恢复可编译。

+++ backend/src/main/java/com/nip/dto/PostTelegramTrainFinishInfoDto.java
@@ -50,7 +50,8 @@ public class PostTelegramTrainFinishInfoDto {
     private Integer key;
 
     // @ApiModelProperty(value = "时长",position = 3)
-    private Integer value;
+    // 时长是客户端单调时钟差值，带小数毫秒；按 Integer 解会让整页结算被判 JSON 损坏。
+    private Double value;
 
   }
 }
